Se modulariza todo los apartado en App
