+++ src/App.jsx
@@ -1,55 +1,14 @@
 import React, { useEffect, useState, useRef } from "react";
 
-// Importación de funcionalidades
-import {
-  foundOrCreateUser,
-  deleteUser,
-  verifyIdUser,
-  changeNameUser,
-  deleteInactiveUsersTwoMonths,
-} from "./data/controllerUsers/controllerUsers";
-
-import {
-  addTaskUser,
-  reviewListTaskUser,
-  readyTaskUser,
-  deleteTaskUser,
-  modifyTaskUser,
-  deleteAllListTaskUser,
-  readyListAllListUser,
-} from "./data/controllerProperties/controllerTasks";
-
-import {
-  addExam,
-  deleteExam,
-  reviewExam,
-  deleteAllExams,
-} from "./data/controllerProperties/controllerExams";
-
-import {
-  getUserInfo,
-  addDataNationality,
-  addBirth,
-  addInstagram,
-  addOppositionfor,
-  addStudyFor,
-  giveCroquetas,
-} from "./data/controllerProperties/controllerPersonalData";
-
 // Importación de cliente de Twitch
 import client from "./data/controllerClientTwitch/clientTwitch.js";
-
-// Importaciones de utilidad con el localStorage
-import jsonData from "./data/localStorageData.json";
-
-import {
-  loadLocalStorageFromFile,
-  saveLocalStorageFile,
-} from "./data/LocalStorage/controllerLocalStorage";
 
 // Importación de componentes
 import UserList from "./components/UserList";
 import InfoUser from "./components/InfoUser";
+
+// Importación de utilidades
+import { monitorMessage } from "./utils/monitorMessage"; // Importamos la función modularizada
 
 const App = () => {
   const [currentUser, setCurrentUser] = useState(null); // Usuario actual mostrado en InfoUser
@@ -62,196 +21,16 @@ const App = () => {
     client.connect();
 
     const handleMessage = (channel, tags, message, self) => {
-      if (self || !message.startsWith("!")) return;
-      const username = tags.username;
-
-      const validCommands = new Set([
-        "id?",
-        "eliminarusuario",
-        "cambiarusuario",
-        "tarea",
-        "add",
-        "task",
-        "lista",
-        "list",
-        "v",
-        "marcar",
-        "check",
-        "x",
-        "eliminar",
-        "borrar",
-        "delete",
-        "modificar",
-        "cambiar",
-        "change",
-        "clear",
-        "borrartodo",
-        "pickup",
-        "realizadas",
-        "nacimiento",
-        "instagram",
-        "opositopara",
-        "estudiopara",
-        "croqueta",
-        "nacionalidad",
-        "datos",
-        "info",
-        "addexam",
-        "examdelete",
-        "reviewexam",
-        "deleteallexam",
-      ]);
-
-      // promps que extraemos del comando
-      const command = message.toLowerCase().split(" ")[0].slice(1);
-      const args = message.slice(1).split(" ");
-      const arg = args[1];
-      const otherUsername = message.slice(7);
-      const taskLowercase = message.substring(command.length + 1);
-      const task =
-        taskLowercase.charAt(0).toUpperCase() + taskLowercase.slice(1);
-      const taskMod =
-        taskLowercase.charAt(0).toUpperCase() + taskLowercase.slice(4);
-
-      const isSub = tags.badges?.subscriber;
-      const isVip = tags.badges?.vip;
-      const isMod = tags.badges?.moderator;
-      const isPrime = tags.badges?.premium;
-
-      const isTag = isSub
-        ? "sub"
-        : isVip
-        ? "vip"
-        : isMod
-        ? "mod"
-        : isPrime
-        ? "prime"
-        : "none";
-
-      // Si el comando no está en la lista, salir
-      if (!validCommands.has(command)) return;
-
-      // Si el mismo usuario ingresa otro comando, extendemos el tiempo de visibilidad
-      if (prevUser.current === username) {
-        clearTimeout(timeoutRef.current); // Cancelamos el temporizador anterior
-        timeoutRef.current = setTimeout(() => {
-          setIsInfoUserVisible(false);
-          setCurrentUser(null);
-        }, 5000); // Solo extendemos 5 segundos más
-      } else {
-        // Si es un nuevo usuario, iniciamos el temporizador estándar
-        prevUser.current = username;
-        clearTimeout(timeoutRef.current);
-        timeoutRef.current = setTimeout(() => {
-          setIsInfoUserVisible(false);
-          setCurrentUser(null);
-        }, 30000);
-      }
-
-      // Manejo de comandos
-      switch (command) {
-        // Administrar usuarios
-        case "id?":
-          verifyIdUser(channel, arg);
-          break;
-        case "eliminarusuario":
-          deleteUser(channel, username, arg);
-          break;
-        case "cambiarusuario":
-          if (username === "cuartodechenz") {
-            changeNameUser(channel, arg, username);
-          }
-          break;
-
-        // Administrar tareas
-        case "tarea":
-        case "add":
-        case "task":
-          addTaskUser(username, task, channel, isTag);
-          break;
-        case "lista":
-        case "list":
-          reviewListTaskUser(username, channel, isTag);
-          break;
-        case "v":
-        case "marcar":
-        case "check":
-          readyTaskUser(username, arg, channel, isTag);
-          break;
-        case "x":
-        case "eliminar":
-        case "borrar":
-        case "delete":
-          deleteTaskUser(username, arg, channel, isTag);
-          break;
-        case "modificar":
-        case "cambiar":
-        case "change":
-          modifyTaskUser(username, arg, taskMod, channel, isTag);
-          break;
-        case "clear":
-        case "borrartodo":
-          deleteAllListTaskUser(username, channel, isTag);
-          break;
-        case "pickup":
-        case "realizadas":
-          readyListAllListUser(username, channel, isTag);
-          break;
-
-        //Administrar información personal de usuarios
-        case "nacimiento":
-          addBirth(username, arg, channel, isTag);
-          break;
-        case "instagram":
-          addInstagram(username, arg, channel);
-          break;
-        case "opositopara":
-          addOppositionfor(username, task, channel);
-          break;
-        case "estudiopara":
-          addStudyFor(username, task, channel);
-          break;
-        case "croqueta":
-          giveCroquetas(username, channel);
-          break;
-        case "nacionalidad":
-          addDataNationality(username, task, channel);
-          break;
-        case "datos":
-          getUserInfo(username, channel);
-          break;
-        case "info":
-          getUserInfo(otherUsername, channel);
-          break;
-
-        // Administrar lista de examenes
-        case "addexam":
-          addExam(username, task, channel, isTag);
-          break;
-        case "examdelete":
-          deleteExam(username, arg, channel, isTag);
-          break;
-        case "reviewexam":
-          reviewExam(username, channel, isTag);
-          break;
-        case "deleteallexam":
-          deleteAllExams(username, channel, isTag);
-          break;
-
-        // Administrar lista de examenes
-        case "guardar":
-          saveLocalStorageFile();
-          break;
-        case "cargar":
-          loadLocalStorageFromFile(jsonData);
-          break;
-
-        default:
-          break;
-      }
-      setCurrentUser(username);
-      setIsInfoUserVisible(true);
-      deleteInactiveUsersTwoMonths();
+      monitorMessage(
+        channel,
+        tags,
+        message,
+        self,
+        prevUser,
+        timeoutRef,
+        setIsInfoUserVisible,
+        setCurrentUser
+      );
     };
 
     client.on("message", handleMessage);
